requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
pandas
numpy
